app/requirements.txt
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
readability-lxml==0.8.1
beautifulsoup4==4.12.3
loguru==0.7.2
tenacity==8.5.0

typer==0.12.3
click==8.1.7
lxml_html_clean>=0.1.0

pandas==2.2.2
numpy==1.26.4
pyarrow==17.0.0
sqlalchemy==2.0.35

psycopg[binary]==3.2.1
pgvector==0.2.5
